--- a/app/package.json
+++ b/app/package.json
@@ -3,7 +3,7 @@
   "productName": "GitHub Desktop",
   "bundleID": "com.github.GitHubClient",
   "companyName": "GitHub, Inc.",
-  "version": "1.1.2-test5",
+  "version": "1.1.2-test6",
   "main": "./main.js",
   "repository": {
     "type": "git",