--- a/app/webpack.common.js
+++ b/app/webpack.common.js
@@ -45,9 +45,10 @@ const commonConfig = {
       },
       {
         test: /\.node$/,
-        use: [
-          { loader: 'node-native-loader', options: { name: '[name].[ext]' } },
-        ],
+        loader: 'awesome-node-loader',
+        options: {
+          name: '[name].[ext]',
+        },
       },
     ],
   },