--- a/app/webpack.production.js
+++ b/app/webpack.production.js
@@ -4,7 +4,7 @@ const common = require('./webpack.common')
 
 const webpack = require('webpack')
 const merge = require('webpack-merge')
-const ExtractTextPlugin = require('extract-text-webpack-plugin')
+const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const BabelPlugin = require('babel-webpack-plugin')
 
 const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer')
@@ -45,16 +45,13 @@ const rendererConfig = merge({}, common.renderer, config, {
       // appended to the index.html HEAD at compile time
       {
         test: /\.(scss|css)$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: ['css-loader', 'sass-loader'],
-        }),
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader'],
       },
     ],
   },
   plugins: [
-    // Necessary to be able to use ExtractTextPlugin as a loader.
-    new ExtractTextPlugin('ui.css'),
+    // Necessary to be able to use MiniCssExtractPlugin as a loader.
+    new MiniCssExtractPlugin('ui.css'),
     new BundleAnalyzerPlugin({
       // this generates the static HTML file to view afterwards, rather
       // than disrupting the user
@@ -76,16 +73,13 @@ const crashConfig = merge({}, common.crash, config, {
       // appended to the index.html HEAD at compile time
       {
         test: /\.(scss|css)$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: ['css-loader', 'sass-loader'],
-        }),
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader'],
       },
     ],
   },
   plugins: [
-    // Necessary to be able to use ExtractTextPlugin as a loader.
-    new ExtractTextPlugin('crash.css'),
+    // Necessary to be able to use MiniCssExtractPlugin as a loader.
+    new MiniCssExtractPlugin('crash.css'),
   ],
 })