--- a/package.json
+++ b/package.json
@@ -56,8 +56,6 @@
     "awesome-typescript-loader": "^3.4.0",
     "aws-sdk": "^2.23.0",
     "babel-core": "^6.26.3",
-    "babel-minify": "^0.4.1",
-    "babel-webpack-plugin": "^0.1.1",
     "chai": "^4.1.1",
     "chai-as-promised": "^7.1.1",
     "chai-datetime": "^1.4.1",