--- a/package.json
+++ b/package.json
@@ -97,7 +97,7 @@
     "tslint-react": "^3.5.1",
     "typescript": "^2.8.1",
     "typescript-eslint-parser": "^15.0.0",
-    "webpack": "^4.7.0",
+    "webpack": "^4.8.1",
     "webpack-bundle-analyzer": "^2.11.1",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-hot-middleware": "^2.22.1",