--- a/package.json
+++ b/package.json
@@ -97,10 +97,10 @@
     "tslint-react": "^3.5.1",
     "typescript": "^2.8.3",
     "typescript-eslint-parser": "^15.0.0",
-    "webpack": "^4.8.1",
-    "webpack-bundle-analyzer": "^2.11.1",
+    "webpack": "^4.8.3",
+    "webpack-bundle-analyzer": "^2.13.0",
     "webpack-dev-middleware": "^3.1.3",
-    "webpack-hot-middleware": "^2.22.1",
+    "webpack-hot-middleware": "^2.22.2",
     "webpack-merge": "^4.1.2",
     "xml2js": "^0.4.16",
     "xvfb-maybe": "^0.2.1"