This is something I've been bumping my head against for a while, and I took another shot at it today with fresh eyes and managed to get it working. I'm mostly opening it at this point to see what improvements are made to build time, and how to address the outstanding CI warnings.

I've been bitten by subtle changes in the compiled output before (see #2531) so I want to be methodical about this. This also overlaps with #3675 somewhat - I'm not sure which I want to land first, to be frank.

 - [x] test `build:dev` output locally
 - [x] `diff` old and new outputs for any differences
 - [x] confirm reloading CSS changes still works
 - [x] test `build:prod` output locally
 - [x] `diff` old and new outputs for any differences
 - [x] investigate performance recommendations emitted by the build and extract tasks
 - [x] publish test build to confirm installers are unaffected
 - [x] tested build on Windows
 - [x] tested build on macOS
 - [x] tested Linux build