@shiftkey 2018-05-04
Welp, it's currently Not Faster At All on AppVeyor - looks like I need to read some migration guides maybe?

Before this PR:

```
[WEBPACK] Finished building ask-pass.js within 24.047 seconds
[WEBPACK] Finished building highlighter.js within 31.735 seconds
[WEBPACK] Finished building cli.js within 46.953 seconds
[WEBPACK] Finished building main.js within 40.313 seconds
[WEBPACK] Finished building crash.js within 58.532 seconds
[WEBPACK] Finished building renderer.js within 128.219 seconds
```

This PR:

```
[WEBPACK] Finished building ask-pass.js within 32.562 seconds
[WEBPACK] Finished building highlighter.js within 46.234 seconds
[WEBPACK] Finished building cli.js within 52.672 seconds
[WEBPACK] Finished building main.js within 52.797 seconds
[WEBPACK] Finished building crash.js within 124.453 seconds
[WEBPACK] Finished building renderer.js within 145.468 seconds
```