@shiftkey 2018-05-04
Nevermind, that last commit I think eliminates a lot of unnecessary work:

```
[WEBPACK] Finished building ask-pass.js within 27.516 seconds
[WEBPACK] Finished building highlighter.js within 41.656 seconds
[WEBPACK] Finished building cli.js within 42.281 seconds
[WEBPACK] Finished building main.js within 42.391 seconds
[WEBPACK] Finished building crash.js within 56.922 seconds
[WEBPACK] Finished building renderer.js within 80.329 seconds
```