@shiftkey 2018-05-08
Another quick benchmark while testing (they're all faster to do the webpack build with `mode: production` set):

|          | [`master`](https://github.com/desktop/desktop/commit/b9a4e82b5108f9b8cd4e06935cbc4b33cd4ef21e) | [`the-great-webpack-migration`](https://github.com/desktop/desktop/pull/4601/commits/26524114bc7edb78af8480a18245f2afdc3b58bc) | Change (%) |
|----------|------------------------------------------------------------------------------------------------|-------------------------------|------------|
| AppVeyor | 119.507s                                                                                       | 78.563s                       | -34%       |
| CircleCI | 141.839s                                                                                       | 69.896s                       | -49%       |
| Travis   | 173.547s                                                                                       | 130.658s                      | -24%       |