@Dinhvamtramg 2018-05-08
[image: Mailtrack]
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality4&>
Sender
notified by
Mailtrack
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality4&>

2018-05-08 11:34 GMT+07:00 Brendan Forster <notifications@github.com>:

> *@shiftkey* commented on this pull request.
> ------------------------------
>
> In app/webpack.production.js
> <https://github.com/desktop/desktop/pull/4601#discussion_r186614679>:
>
> >    devtool: 'source-map',
> -  plugins: [
>
> Digging into this some more, it looks like uglify-es
> <https://www.npmjs.com/package/uglify-es> is the equivalent ES6+ support
> for UglifyJS, and after going down the path of installing
> uglify-es-webpack-plugin I instead got this warning/recommendation:
>
> warning uglify-es-webpack-plugin@0.10.0: webpack-contrib/uglifyjs-webpack-plugin now has uglify-es support, featuring tests and far better implementation
>
> The version we now have inside yarn.lock also contains uglify-es:
>
> uglifyjs-webpack-plugin@^1.2.4:
>   version "1.2.5"
>   resolved "https://registry.yarnpkg.com/uglifyjs-webpack-plugin/-/uglifyjs-webpack-plugin-1.2.5.tgz#2ef8387c8f1a903ec5e44fa36f9f3cbdcea67641"
>   dependencies:
>     cacache "^10.0.4"
>     find-cache-dir "^1.0.0"
>     schema-utils "^0.4.5"
>     serialize-javascript "^1.4.0"
>     source-map "^0.6.1"
>     uglify-es "^3.3.4"
>     webpack-sources "^1.1.0"
>     worker-farm "^1.5.2"
>
> So I don't really see the value of switching back to babel-minify if
> Uglify has caught up enough and we still have the tree-shaking and dead
> code elimination that motivated #1225
> <https://github.com/desktop/desktop/pull/1225> in the first place.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/desktop/desktop/pull/4601#discussion_r186614679>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AjnZ25bPB_smmLonFFvngpR0fr04YxEUks5twSA7gaJpZM4TyKaD>
> .
>