@niik 2018-05-09
@shiftkey I'm trying something out based on this work and I found that I needed to bump awesome-typescript-loader to get rid of a crash on recompilation when running in the devserver. I pushed that change straight to this branch.

> confirm reloading CSS changes still works - not working, see s-panferov/awesome-typescript-loader#534

I can't reproduce this but I don't know if it was solved by the bump or not. Could you see if it's still not working for you?