@shiftkey 2018-05-09
@niik yeah, I was testing pre-release packages for `awesome-typescript-loader` to tackle that problem, and I think `5.0.0-*` was close to addressing the issue. I'm not sure why I put it aside, but I'll give it another shot today and see if I can still trip the issue.